requirements.txt
click
loguru
protobuf